+++ tests/test_python_script.py
@@ -54,8 +54,9 @@ def test_write_build_python_script(tmpdir):
     with open(os.path.join(tmpdir, "foo.py"), "w") as fh:
         fh.write(SCRIPT)
     with contextlib.chdir(tmpdir):
-        builds, _ = python_script.generate(["foo.py"], [], None, {})
+        builds, ns = python_script.generate(["foo.py"], [], None, {})
     assert BUILDS_PYTHON_SCRIPT == builds
+    assert ns == ["foo.py"]
 
 
 BUILDS_PYTHON_SCRIPT_SUB = [
@@ -78,5 +79,6 @@ def test_write_build_python_script_sub(tmpdir):
     with open(os.path.join(subdir, "foo.py"), "w") as fh:
         fh.write(SCRIPT)
     with contextlib.chdir(tmpdir):
-        builds, _ = python_script.generate(["sub/foo.py"], [], None, {})
+        builds, ns = python_script.generate(["sub/foo.py"], [], None, {})
     assert BUILDS_PYTHON_SCRIPT_SUB == builds
+    assert ns == ["sub/foo.py"]